vfio-user: support for shadow ioeventfd
This patch enables QEMU to use the shadow ioeventfd mechanism. This
mechanism avoids an exit to the VMM when a GPA is written to and the
value being written is needed by the device emulation task (e.g. an SQ
doorbell write of an NVMe controller). It works by setting up an
ioeventfd except that whenever the GPA is written to, KVM writes the
otherwise discarded value to a piece of memory visible only to KVM/VMM
and the device emulation task.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>